requirements.txt
psycopg2-binary==2.8.6
aiogram~=2.13
SQLAlchemy~=1.4.15
pandas~=1.2.4
emoji==1.2.0